manamoa/black_devil_king/test.html

<!--<![endif]--><head>
<title>Attention Required! | Cloudflare</title>
<meta name="captcha-bypass" id="captcha-bypass">
<meta charset="UTF-8">
<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
<meta http-equiv="X-UA-Compatible" content="IE=Edge,chrome=1">
<meta name="robots" content="noindex, nofollow">
<meta name="viewport" content="width=device-width,initial-scale=1,maximum-scale=1">
<link rel="stylesheet" id="cf_styles-css" href="/cdn-cgi/styles/cf.errors.css" type="text/css" media="screen,projection">
<!--[if lt IE 9]><link rel="stylesheet" id='cf_styles-ie-css' href="/cdn-cgi/styles/cf.errors.ie.css" type="text/css" media="screen,projection" /><![endif]-->
<style type="text/css">body{margin:0;padding:0}</style>


<!--[if gte IE 10]><!--><script type="text/javascript" async="" src="https://www.gstatic.com/recaptcha/releases/66WEle60vY1w2WveBS-1ZMFs/recaptcha__en.js"></script><script type="text/javascript" src="/cdn-cgi/scripts/zepto.min.js"></script><!--<![endif]-->
<!--[if gte IE 10]><!--><script type="text/javascript" src="/cdn-cgi/scripts/cf.common.js"></script><!--<![endif]-->




<script charset="utf8" src="https://www.google.com/recaptcha/api.js?onload=onloadCallback&amp;render=explicit"></script></head>
<body>
  <div id="cf-wrapper">
    <div class="cf-alert cf-alert-error cf-cookie-error" id="cookie-alert" data-translate="enable_cookies">Please enable cookies.</div>
    <div id="cf-error-details" class="cf-error-details-wrapper">
      <div class="cf-wrapper cf-header cf-error-overview">
        <h1 data-translate="challenge_headline">One more step</h1>
        <h2 class="cf-subheadline"><span data-translate="complete_sec_check">Please complete the security check to access</span> manamoa17.net</h2>
      </div><!-- /.header -->
      
      <div class="cf-section cf-highlight cf-captcha-container">
        <div class="cf-wrapper">
          <div class="cf-columns two">
            <div class="cf-column">
            
              <div class="cf-highlight-inverse cf-form-stacked">
                <form class="challenge-form" id="challenge-form" action="/cdn-cgi/l/chk_captcha" method="get">
  <input type="hidden" name="s" value="d43c6099ff007b916138de6c9cfd57677d31e197-1572832701-0-AVyBtA8LihouW/uObvgjrKK/laNweHNqQeQF+WFGJJW4/o2JomH0Q9tuJvCY1TY4jiQVRLqN3Uw2HhUJ7ZB2bSTlskymKmY562QiqkbenC9oasleXBUyz8MeTCjpyTsa3mfHJeGUjAtaaVzPQ792VQu7E7rTmK9zybTLNTf5AkPRtxJUQC73w3sohbGtmUkow6kA/HqYu2lZDnNEhtFG03bCSa/wG2c3yC1V0dCwoWaBhVSvVH11rnehNs319y0FOQ==">
  <script type="text/javascript" src="/cdn-cgi/scripts/cf.challenge.js" data-type="normal" data-ray="5302fb801a86dd7b" async="" data-sitekey="6LfBixYUAAAAABhdHynFUIMA_sa4s-XsJvnjtgB0"></script>
  <div class="g-recaptcha"></div>
  <input id="id" name="id" type="hidden" value="5302fb801a86dd7b"> <div id="recaptcha_widget"><div style="width: 304px; height: 78px;"><div><iframe src="https://www.google.com/recaptcha/api2/anchor?ar=1&amp;k=6LfBixYUAAAAABhdHynFUIMA_sa4s-XsJvnjtgB0&amp;co=aHR0cHM6Ly9tYW5hbW9hMTcubmV0OjQ0Mw..&amp;hl=en&amp;v=66WEle60vY1w2WveBS-1ZMFs&amp;size=normal&amp;cb=a11p2jf1z6ii" width="304" height="78" role="presentation" name="a-3kxsrf2lla77" frameborder="0" scrolling="no" sandbox="allow-forms allow-popups allow-same-origin allow-scripts allow-top-navigation allow-modals allow-popups-to-escape-sandbox"></iframe></div><textarea id="g-recaptcha-response" name="g-recaptcha-response" class="g-recaptcha-response" style="width: 250px; height: 40px; border: 1px solid rgb(193, 193, 193); margin: 10px 25px; padding: 0px; resize: none; display: none;"></textarea></div></div> <input id="recaptcha_submit" type="submit" value="Submit"> <noscript id="cf-captcha-bookmark" class="cf-captcha-info">
    <div><div style="width: 302px">
      <div>
        <iframe src="https://www.google.com/recaptcha/api/fallback?k=6LfBixYUAAAAABhdHynFUIMA_sa4s-XsJvnjtgB0" frameborder="0" scrolling="no" style="width: 302px; height:422px; border-style: none;"></iframe>
      </div>
      <div style="width: 300px; border-style: none; bottom: 12px; left: 25px; margin: 0px; padding: 0px; right: 25px; background: #f9f9f9; border: 1px solid #c1c1c1; border-radius: 3px;">
        <textarea id="g-recaptcha-response" name="g-recaptcha-response" class="g-recaptcha-response" style="width: 250px; height: 40px; border: 1px solid #c1c1c1; margin: 10px 25px; padding: 0px; resize: none;"></textarea>
        <input type="submit" value="Submit"></input>
      </div>
    </div></div>
  </noscript>
<input type="hidden" name="cv_chal_result" value="[&quot;96e4643ff97653b1f5f67e3cc9a1a7c0&quot;,&quot;78ac47a81ebde0bce87bb7eee0196a3d&quot;]"><input type="hidden" name="cv_chal_fp" value="{&quot;r&quot;:[1920,1080],&quot;ar&quot;:[1080,1920],&quot;pr&quot;:1,&quot;cd&quot;:24}"><input type="hidden" name="bf_challenge_id" value="8381"><input type="hidden" name="bf_execution_time" value="93"><input type="hidden" name="bf_result_hash" value="3260869"></form>

                <script type="text/javascript">
  (function(){
    var a = function() {try{return !!window.addEventListener} catch(e) {return !1} },
    b = function(b, c) {a() ? document.addEventListener("DOMContentLoaded", b, c) : document.attachEvent("onreadystatechange", b)};
    b(function(){
      if (!a()) return;

      window.addEventListener("message", handleMessage, false)

      function handleMessage(event) {
        if (event.data && event.data.type === 'results') {
          var f = document.getElementById('challenge-form');

          if (f) {
            addInput(f, 'bf_challenge_id', '8381');
            addInput(f, 'bf_execution_time', event.data.executionTimeMs);
            addInput(f, 'bf_result_hash', event.data.resultHash);
          }

          window.removeEventListener("message", handleMessage, false)
        }
      }

      function addInput(parent, name, value) {
        var input = document.createElement('input');
        input.type = 'hidden';
        input.name = name;
        input.value = value;
        parent.appendChild(input);
      }

      function withIframe(iframeContent) {
        var iframe = document.createElement('iframe');
        iframe.id = 'bf_test_iframe';
        iframe.style.visibility = 'hidden';
        document.body.appendChild(iframe);
        var doc = (iframe.contentWindow || iframe.contentDocument).document;
        doc.write(iframeContent);
        doc.close();
      }

      withIframe("<!DOCTYPE html>\n<html>\n  <head>\n    <title><\/title>\n    <script src=\"https:\/\/ajax.cloudflare.com\/cdn-cgi\/scripts\/697236fc\/cloudflare-static\/bot-filter.js\"><\/__script__>\n    \n    \n    \n    \n  <\/head>\n  <body>\n    <script id=\"layout-test-code\">var RENDER_QUANTUM_FRAMES=128;function compareBuffersWithConstraints(should,actual,expected,options){options||(options={}),actual.length!==expected.length&&should(actual.length===expected.length,\"Length of actual and expected buffers should match\").beTrue();for(var maxError=-1,diffCount=0,errorPosition=-1,thresholdSNR=options.thresholdSNR||1e4,thresholdDiffULP=options.thresholdDiffULP||0,thresholdDiffCount=options.thresholdDiffCount||0,bitDepth=options.bitDepth||16,scaleFactor=Math.pow(2,bitDepth-1),noisePower=0,signalPower=0,i=0;i<actual.length;i++){var diff=actual[i]-expected[i];noisePower+=diff*diff,signalPower+=expected[i]*expected[i],Math.abs(diff)>maxError&&(maxError=Math.abs(diff),errorPosition=i),Math.abs(diff)>scaleFactor&&diffCount++}var snr,maxErrorULP=maxError*scaleFactor;should(10*Math.log10(signalPower\/noisePower),\"SNR\").beGreaterThanOrEqualTo(thresholdSNR),should(maxErrorULP,options.prefix+\": Maximum difference (in ulp units (\"+bitDepth+\"-bits))\").beLessThanOrEqualTo(thresholdDiffULP),should(diffCount,options.prefix+\": Number of differences between results\").beLessThanOrEqualTo(thresholdDiffCount)}function createImpulseBuffer(context,sampleFrameLength){for(var audioBuffer=context.createBuffer(1,sampleFrameLength,context.sampleRate),n=audioBuffer.length,dataL=audioBuffer.getChannelData(0),k=0;k<n;++k)dataL[k]=0;return dataL[0]=1,audioBuffer}function createLinearRampBuffer(context,sampleFrameLength){for(var audioBuffer=context.createBuffer(1,sampleFrameLength,context.sampleRate),n=audioBuffer.length,dataL=audioBuffer.getChannelData(0),i=0;i<n;++i)dataL[i]=i\/n;return audioBuffer}function createConstantBuffer(context,sampleFrameLength,constantValue){var channels,values;\"number\"==typeof constantValue?(channels=1,values=[constantValue]):(channels=constantValue.length,values=constantValue);for(var audioBuffer=context.createBuffer(channels,sampleFrameLength,context.sampleRate),n=audioBuffer.length,c=0;c<channels;++c)for(var data=audioBuffer.getChannelData(c),i=0;i<n;++i)data[i]=values[c];return audioBuffer}function createStereoImpulseBuffer(context,sampleFrameLength){for(var audioBuffer=context.createBuffer(2,sampleFrameLength,context.sampleRate),n=audioBuffer.length,dataL=audioBuffer.getChannelData(0),dataR=audioBuffer.getChannelData(1),k=0;k<n;++k)dataL[k]=0,dataR[k]=0;return dataL[0]=1,dataR[0]=1,audioBuffer}function timeToSampleFrame(time,sampleRate){return Math.floor(.5+time*sampleRate)}function grainLengthInSampleFrames(grainOffset,duration,sampleRate){var startFrame=timeToSampleFrame(grainOffset,sampleRate),endFrame;return timeToSampleFrame(grainOffset+duration,sampleRate)-startFrame}function isValidNumber(x){return!isNaN(x)&&x!=1\/0&&x!=-1\/0}function computeSNR(actual,expected){for(var signalPower=0,noisePower=0,length=Math.min(actual.length,expected.length),k=0;k<length;++k){var diff=actual[k]-expected[k];signalPower+=expected[k]*expected[k],noisePower+=diff*diff}return signalPower\/noisePower}function _classCallCheck(instance,Constructor){if(!(instance instanceof Constructor))throw new TypeError(\"Cannot call a class as a function\")}function _defineProperties(target,props){for(var i=0;i<props.length;i++){var descriptor=props[i];descriptor.enumerable=descriptor.enumerable||!1,descriptor.configurable=!0,\"value\"in descriptor&&(descriptor.writable=!0),Object.defineProperty(target,descriptor.key,descriptor)}}function _createClass(Constructor,protoProps,staticProps){return protoProps&&_defineProperties(Constructor.prototype,protoProps),staticProps&&_defineProperties(Constructor,staticProps),Constructor}function _typeof(obj){return(_typeof=\"function\"==typeof Symbol&&\"symbol\"==typeof Symbol.iterator?function _typeof(obj){return typeof obj}:function _typeof(obj){return obj&&\"function\"==typeof Symbol&&obj.constructor===Symbol&&obj!==Symbol.prototype?\"symbol\":typeof obj})(obj)}function createLowpassFilter(freq,q,gain){var b0,b1,b2,a0,a1,a2;if(1==freq)a0=b0=1,a2=a1=b2=b1=0;else{var theta=Math.PI*freq,alpha=Math.sin(theta)\/(2*Math.pow(10,q\/20)),cosw=Math.cos(theta),beta=(1-cosw)\/2;b1=2*(b0=beta),b2=beta,a0=1+alpha,a1=-2*cosw,a2=1-alpha}return normalizeFilterCoefficients(b0,b1,b2,a0,a1,a2)}function createHighpassFilter(freq,q,gain){var b0,b1,b2,a0,a1,a2;if(1==freq)a0=1,a2=a1=b2=b1=b0=0;else if(0==freq)a0=b0=1,a2=a1=b2=b1=0;else{var theta=Math.PI*freq,alpha=Math.sin(theta)\/(2*Math.pow(10,q\/20)),cosw=Math.cos(theta),beta=(1+cosw)\/2;b1=-2*(b0=beta),b2=beta,a0=1+alpha,a1=-2*cosw,a2=1-alpha}return normalizeFilterCoefficients(b0,b1,b2,a0,a1,a2)}function normalizeFilterCoefficients(b0,b1,b2,a0,a1,a2){var scale=1\/a0;return{b0:b0*scale,b1:b1*scale,b2:b2*scale,a1:a1*scale,a2:a2*scale}}function createBandpassFilter(freq,q,gain){var b0,b1,b2,a0,a1,a2,coef;if(0<freq&&freq<1){var w0=Math.PI*freq;if(0<q){var alpha=Math.sin(w0)\/(2*q),k;coef=normalizeFilterCoefficients(b0=alpha,b1=0,b2=-alpha,a0=1+alpha,a1=-2*Math.cos(w0),a2=1-alpha)}else coef={b0:1,b1:0,b2:0,a1:0,a2:0}}else coef={b0:0,b1:0,b2:0,a1:0,a2:0};return coef}function createLowShelfFilter(freq,q,gain){var b0,b1,b2,a0,a1,a2,coef,S=1,A=Math.pow(10,gain\/40);if(1==freq)coef={b0:A*A,b1:0,b2:0,a1:0,a2:0};else if(0==freq)coef={b0:1,b1:0,b2:0,a1:0,a2:0};else{var w0=Math.PI*freq,alpha=.5*Math.sin(w0)*Math.sqrt(0*(A+1\/A)+2),k=Math.cos(w0),k2=2*Math.sqrt(A)*alpha,Ap1=A+1,Am1=A-1;coef=normalizeFilterCoefficients(b0=A*(Ap1-Am1*k+k2),b1=2*A*(Am1-Ap1*k),b2=A*(Ap1-Am1*k-k2),a0=Ap1+Am1*k+k2,a1=-2*(Am1+Ap1*k),a2=Ap1+Am1*k-k2)}return coef}function createHighShelfFilter(freq,q,gain){var b0,b1,b2,a0,a1,a2,coef,A=Math.pow(10,gain\/40);if(1==freq)coef={b0:1,b1:0,b2:0,a1:0,a2:0};else if(0<freq){var w0=Math.PI*freq,S=1,alpha=.5*Math.sin(w0)*Math.sqrt(0*(A+1\/A)+2),k=Math.cos(w0),k2=2*Math.sqrt(A)*alpha,Ap1=A+1,Am1=A-1;coef=normalizeFilterCoefficients(b0=A*(Ap1+Am1*k+k2),b1=-2*A*(Am1+Ap1*k),b2=A*(Ap1+Am1*k-k2),a0=Ap1-Am1*k+k2,a1=2*(Am1-Ap1*k),a2=Ap1-Am1*k-k2)}else coef={b0:A*A,b1:0,b2:0,a1:0,a2:0};return coef}function createPeakingFilter(freq,q,gain){var b0,b1,b2,a0,a1,a2,coef,A=Math.pow(10,gain\/40);if(0<freq&&freq<1)if(0<q){var w0=Math.PI*freq,alpha=Math.sin(w0)\/(2*q),k=Math.cos(w0);coef=normalizeFilterCoefficients(b0=1+alpha*A,b1=-2*k,b2=1-alpha*A,a0=1+alpha\/A,a1=-2*k,a2=1-alpha\/A)}else coef={b0:A*A,b1:0,b2:0,a1:0,a2:0};else coef={b0:1,b1:0,b2:0,a1:0,a2:0};return coef}function createNotchFilter(freq,q,gain){var b0,b1,b2,a0,a1,a2,coef;if(0<freq&&freq<1)if(0<q){var w0=Math.PI*freq,alpha=Math.sin(w0)\/(2*q),k=Math.cos(w0);coef=normalizeFilterCoefficients(b2=b0=1,b1=-2*k,1,a0=1+alpha,a1=-2*k,a2=1-alpha)}else coef={b0:0,b1:0,b2:0,a1:0,a2:0};else coef={b0:1,b1:0,b2:0,a1:0,a2:0};return coef}function createAllpassFilter(freq,q,gain){var b0,b1,b2,a0,a1,a2,coef;if(0<freq&&freq<1)if(0<q){var w0=Math.PI*freq,alpha=Math.sin(w0)\/(2*q),k=Math.cos(w0);coef=normalizeFilterCoefficients(b0=1-alpha,b1=-2*k,b2=1+alpha,a0=1+alpha,a1=-2*k,a2=1-alpha)}else coef={b0:-1,b1:0,b2:0,a1:0,a2:0};else coef={b0:1,b1:0,b2:0,a1:0,a2:0};return coef}function filterData(filterCoef,signal,len){var y=new Array(len),b0=filterCoef.b0,b1=filterCoef.b1,b2=filterCoef.b2,a1=filterCoef.a1,a2=filterCoef.a2;y[0]=b0*signal[0],y[1]=b0*signal[1]+b1*signal[0]-a1*y[0];for(var k=2;k<Math.min(signal.length,len);++k)y[k]=b0*signal[k]+b1*signal[k-1]+b2*signal[k-2]-a1*y[k-1]-a2*y[k-2];for(var _k=signal.length;_k<len;++_k)y[_k]=-a1*y[_k-1]-a2*y[_k-2];return y}!function(){\"use strict\";var testharnessProperties=[\"test\",\"async_test\",\"promise_test\",\"promise_rejects\",\"generate_tests\",\"setup\",\"done\",\"__c$1\",\"__c$1\"];for(var name in testharnessProperties)if(!self.hasOwnProperty(testharnessProperties[name]))throw new Error(\"Cannot proceed. testharness.js is not loaded.\")}(),window.Audit=function(){\"use strict\";function _logPassed(message){test(function(arg){__c$1(!0)})}function _logFailed(message,detail){}function _throwException(message){throw new Error(message)}var _testharnessDone=window.done;function _generateDescription(target,options){var targetString;switch(_typeof(target)){case\"object\":var arrayElements;if(target instanceof Array||target instanceof Float32Array||target instanceof Float64Array||target instanceof Uint8Array)targetString=\"[\"+(target.length<options.numberOfArrayElements?String(target):String(target.slice(0,options.numberOfArrayElements))+\"...\")+\"]\";else targetString=null===target?String(target):\"\"+String(targetString).split(\/[\\s\\]]\/)[1];break;default:targetString=String(target);break}return targetString}function _formatFailureEntry(index,actual,expected,abserr,threshold){return\"\\t[\"+index+\"]\\t\"+actual.toExponential(16)+\"\\t\"+expected.toExponential(16)+\"\\t\"+abserr.toExponential(16)+\"\\t\"+(abserr\/Math.abs(expected)).toExponential(16)+\"\\t\"+threshold.toExponential(16)}function _closeToThreshold(abserr,relerr,expected){return Math.max(abserr,relerr*Math.abs(expected))}window.done=function(){_throwException(\"Do NOT call done() method from the test code.\")};var Should=function(){function Should(parentTask,actual,actualDescription){_classCallCheck(this,Should),this._task=parentTask,this._actual=actual,this._actualDescription=actualDescription||null,this._expected=null,this._expectedDescription=null,this._detail=\"\",this._printActualForFailure=!0,this._result=null,this._options={numberOfErrors:4,numberOfArrayElements:16,verbose:!1}}return _createClass(Should,[{key:\"_processArguments\",value:function _processArguments(args){0!==args.length&&(0<args.length&&(this._expected=args[0]),\"string\"==typeof args[1]?(this._expectedDescription=args[1],Object.assign(this._options,args[2])):\"object\"===_typeof(args[1])&&Object.assign(this._options,args[1]))}},{key:\"_buildResultText\",value:function _buildResultText(){null===this._result&&_throwException(\"Illegal invocation: the assertion is not finished.\");var actualString=_generateDescription(this._actual,this._options);for(var name in this._actualDescription||(this._actualDescription=actualString),this._expectedDescription||(this._expectedDescription=_generateDescription(this._expected,this._options)),this._detail=this._detail.replace(\/\\$\\{actual\\}\/g,this._actualDescription),this._detail=this._detail.replace(\/\\$\\{expected\\}\/g,this._expectedDescription),this._options)if(\"numberOfErrors\"!==name&&\"numberOfArrayElements\"!==name&&\"verbose\"!==name){var re=\"${\"+name+\"}\";re=re.replace(\/([.*+?^=!:${}()|\\[\\]\\\/\\\\])\/g,\"\\\\$1\"),this._detail=this._detail.replace(new RegExp(re,\"g\"),_generateDescription(this._options[name]))}!1===this._result&&!0===this._printActualForFailure&&(this._detail+=\" Got \"+actualString+\".\")}},{key:\"_finalize\",value:function _finalize(){this._result?_logPassed(\"  \"+this._detail):this._detail,this._task.update(this)}},{key:\"_assert\",value:function _assert(condition,passDetail,failDetail){return this._result=Boolean(condition),this._detail=this._result?passDetail:failDetail,this._buildResultText(),this._finalize(),this._result}},{key:\"exist\",value:function exist(){return this._assert(null!==this._actual&&void 0!==this._actual,\"${actual} does exist.\",\"${actual} does not exist.\")}},{key:\"throw\",value:function _throw(){this._processArguments(arguments);var didThrowCorrectly=this._printActualForFailure=!1,passDetail,failDetail;try{this._actual(),failDetail=\"${actual} did not throw an exception.\"}catch(error){var errorMessage=this._options.omitErrorMessage?\": [error message omitted]\":': \"'+error.message+'\"';null===this._expected||void 0===this._expected?(didThrowCorrectly=!0,passDetail=\"${actual} threw \"+error.name+errorMessage+\".\"):error.name===this._expected?(didThrowCorrectly=!0,passDetail=\"${actual} threw ${expected}\"+errorMessage+\".\"):(didThrowCorrectly=!1,failDetail='${actual} threw \"'+error.name+'\" instead of ${expected}.')}return this._assert(didThrowCorrectly,passDetail,failDetail)}},{key:\"notThrow\",value:function notThrow(){var didThrowCorrectly=this._printActualForFailure=!1,passDetail,failDetail;try{this._actual(),passDetail=\"${actual} did not throw an exception.\"}catch(error){didThrowCorrectly=!0,failDetail=\"${actual} incorrectly threw \"+error.name+': \"'+error.message+'\".'}return this._assert(!didThrowCorrectly,passDetail,failDetail)}},{key:\"beResolved\",value:function beResolved(){return this._actual.then(function(result){return this._assert(!0,\"${actual} resolved correctly.\",null),result}.bind(this),function(error){this._assert(!1,null,\"${actual} rejected incorrectly with \"+error+\".\")}.bind(this))}},{key:\"beRejected\",value:function beRejected(){return this._actual.then(function(){this._assert(!1,null,\"${actual} resolved incorrectly.\")}.bind(this),function(error){this._assert(!0,\"${actual} rejected correctly with \"+error+\".\",null)}.bind(this))}},{key:\"beRejectedWith\",value:function beRejectedWith(){return this._processArguments(arguments),this._actual.then(function(){this._assert(!1,null,\"${actual} resolved incorrectly.\")}.bind(this),function(error){this._expected!==error.name?this._assert(!1,null,\"${actual} rejected correctly but got \"+error.name+\" instead of \"+this._expected+\".\"):this._assert(!0,\"${actual} rejected correctly with \"+this._expected+\".\",null)}.bind(this))}},{key:\"beTrue\",value:function beTrue(){return this._assert(!0===this._actual,\"${actual} is true.\",\"${actual} is not true.\")}},{key:\"beFalse\",value:function beFalse(){return this._assert(!1===this._actual,\"${actual} is false.\",\"${actual} is not false.\")}},{key:\"beEqualTo\",value:function beEqualTo(){return this._processArguments(arguments),this._assert(this._actual===this._expected,\"${actual} is equal to ${expected}.\",\"${actual} is not equal to ${expected}.\")}},{key:\"notBeEqualTo\",value:function notBeEqualTo(){return this._processArguments(arguments),this._assert(this._actual!==this._expected,\"${actual} is not equal to ${expected}.\",\"${actual} should not be equal to ${expected}.\")}},{key:\"beNaN\",value:function beNaN(){return this._processArguments(arguments),this._assert(isNaN(this._actual),\"${actual} is NaN.\",\"${actual} is not NaN but should be.\")}},{key:\"notBeNaN\",value:function notBeNaN(){return this._processArguments(arguments),this._assert(!isNaN(this._actual),\"${actual} is not NaN.\",\"${actual} is NaN but should not be.\")}},{key:\"beGreaterThan\",value:function beGreaterThan(){return this._processArguments(arguments),this._assert(this._actual>this._expected,\"${actual} is greater than ${expected}.\",\"${actual} is not greater than ${expected}.\")}},{key:\"beGreaterThanOrEqualTo\",value:function beGreaterThanOrEqualTo(){return this._processArguments(arguments),this._assert(this._actual>=this._expected,\"${actual} is greater than or equal to ${expected}.\",\"${actual} is not greater than or equal to ${expected}.\")}},{key:\"beLessThan\",value:function beLessThan(){return this._processArguments(arguments),this._assert(this._actual<this._expected,\"${actual} is less than ${expected}.\",\"${actual} is not less than ${expected}.\")}},{key:\"beLessThanOrEqualTo\",value:function beLessThanOrEqualTo(){return this._processArguments(arguments),this._assert(this._actual<=this._expected,\"${actual} is less than or equal to ${expected}.\",\"${actual} is not less than or equal to ${expected}.\")}},{key:\"beConstantValueOf\",value:function beConstantValueOf(){this._processArguments(arguments);for(var passed=!(this._printActualForFailure=!1),passDetail,failDetail,errors={},actual=this._actual,expected=this._expected,index=0;index<actual.length;++index)actual[index]!==expected&&(errors[index]=actual[index]);var numberOfErrors=Object.keys(errors).length;if(passed=0===numberOfErrors)passDetail=\"${actual} contains only the constant ${expected}.\";else{var counter=0;for(var errorIndex in failDetail=\"${actual}: Expected ${expected} for all values but found \"+numberOfErrors+\" unexpected values: \",failDetail+=\"\\n\\tIndex\\tActual\",errors)if(failDetail+=\"\\n\\t[\"+errorIndex+\"]\\t\"+errors[errorIndex],++counter>=this._options.numberOfErrors){failDetail+=\"\\n\\t...and \"+(numberOfErrors-counter)+\" more errors.\";break}}return this._assert(passed,passDetail,failDetail)}},{key:\"notBeConstantValueOf\",value:function notBeConstantValueOf(){this._processArguments(arguments);for(var passed=!(this._printActualForFailure=!1),passDetail,failDetail,differences={},actual=this._actual,expected=this._expected,index=0;index<actual.length;++index)actual[index]!==expected&&(differences[index]=actual[index]);var numberOfDifferences=Object.keys(differences).length,valueString;(passed=0<numberOfDifferences)?passDetail=\"${actual} is not constantly ${expected} (contains \"+numberOfDifferences+\" different \"+(1<numberOfDifferences?\"values\":\"value\")+\").\":failDetail=\"${actual} should have contain at least one value different from ${expected}.\";return this._assert(passed,passDetail,failDetail)}},{key:\"beEqualToArray\",value:function beEqualToArray(){this._processArguments(arguments);var passed=!(this._printActualForFailure=!1),passDetail,failDetail,errorIndices=[];if(this._actual.length!==this._expected.length)return passed=!1,failDetail=\"The array length does not match.\",this._assert(passed,passDetail,failDetail);for(var actual=this._actual,expected=this._expected,index=0;index<actual.length;++index)actual[index]!==expected[index]&&errorIndices.push(index);if(passed=0===errorIndices.length)passDetail=\"${actual} is identical to the array ${expected}.\";else{var counter=0;failDetail=\"${actual} expected to be equal to the array ${expected} but differs in \"+errorIndices.length+\" places:\\n\\tIndex\\tActual\\t\\t\\tExpected\";for(var _i=0;_i<errorIndices.length;_i++){var _index=errorIndices[_i];if(failDetail+=\"\\n\\t[\"+_index+\"]\\t\"+this._actual[_index].toExponential(16)+\"\\t\"+this._expected[_index].toExponential(16),++counter>=this._options.numberOfErrors){failDetail+=\"\\n\\t...and \"+(errorIndices.length-counter)+\" more errors.\";break}}}return this._assert(passed,passDetail,failDetail)}},{key:\"containValues\",value:function containValues(){this._processArguments(arguments);for(var passed=!(this._printActualForFailure=!1),indexedActual=[],firstErrorIndex=null,i=0,prev=null;i<this._actual.length;i++)this._actual[i]!==prev&&(indexedActual.push({index:i,value:this._actual[i]}),prev=this._actual[i]);for(var j=0;j<this._expected.length;j++)if(this._expected[j]!==indexedActual[j].value){firstErrorIndex=indexedActual[j].index,passed=!1;break}return this._assert(passed,\"${actual} contains all the expected values in the correct order: ${expected}.\",\"${actual} expected to have the value sequence of ${expected} but got \"+this._actual[firstErrorIndex]+\" at index \"+firstErrorIndex+\".\")}},{key:\"notGlitch\",value:function notGlitch(){this._processArguments(arguments);for(var passed=!(this._printActualForFailure=!1),passDetail,failDetail,actual=this._actual,expected=this._expected,index=0;index<actual.length;++index){var diff=Math.abs(actual[index-1]-actual[index]);expected<=diff&&(passed=!1,failDetail=\"${actual} has a glitch at index \"+index+\" of size \"+diff+\".\")}return passDetail=\"${actual} has no glitch above the threshold of ${expected}.\",this._assert(passed,passDetail,failDetail)}},{key:\"beCloseTo\",value:function beCloseTo(){this._processArguments(arguments);var absExpected=this._expected?Math.abs(this._expected):1,error=Math.abs(this._actual-this._expected)\/absExpected;return this._assert(error<=this._options.threshold,\"${actual} is ${expected} within an error of ${threshold}.\",\"${actual} is not close to ${expected} within a relative error of ${threshold} (RelErr=\"+error+\").\")}},{key:\"beCloseToArray\",value:function beCloseToArray(){this._processArguments(arguments);for(var passed=!(this._printActualForFailure=!1),passDetail,failDetail,absErrorThreshold=this._options.absoluteThreshold||0,relErrorThreshold=this._options.relativeThreshold||0,errors={},maxAbsError=-1\/0,maxAbsErrorIndex=-1,maxRelError=-1\/0,maxRelErrorIndex=-1,actual=this._actual,expected=this._expected,index=0;index<expected.length;++index){var diff=Math.abs(actual[index]-expected[index]),absExpected=Math.abs(expected[index]),relError=diff\/absExpected;diff>Math.max(absErrorThreshold,relErrorThreshold*absExpected)&&(maxAbsError<diff&&(maxAbsErrorIndex=index,maxAbsError=diff),!isNaN(relError)&&maxRelError<relError&&(maxRelErrorIndex=index,maxRelError=relError),errors[index]=diff)}var numberOfErrors=Object.keys(errors).length,maxAllowedErrorDetail=JSON.stringify({absoluteThreshold:absErrorThreshold,relativeThreshold:relErrorThreshold});if(0===numberOfErrors)passDetail=\"${actual} equals ${expected} with an element-wise tolerance of \"+maxAllowedErrorDetail+\".\";else{passed=!1,failDetail=\"${actual} does not equal ${expected} with an element-wise tolerance of \"+maxAllowedErrorDetail+\".\\n\";var counter=0;failDetail+=\"\\tIndex\\tActual\\t\\t\\tExpected\\t\\tAbsError\\t\\tRelError\\t\\tTest threshold\";var printedIndices=[];for(var _index2 in errors)if(failDetail+=\"\\n\"+_formatFailureEntry(_index2,actual[_index2],expected[_index2],errors[_index2],_closeToThreshold(absErrorThreshold,relErrorThreshold,expected[_index2])),printedIndices.push(_index2),++counter>this._options.numberOfErrors){failDetail+=\"\\n\\t...and \"+(numberOfErrors-counter)+\" more errors.\";break}failDetail+=\"\\n\\tMax AbsError of \"+maxAbsError.toExponential(16)+\" at index of \"+maxAbsErrorIndex+\".\\n\",void 0===printedIndices.find(function(element){return element==maxAbsErrorIndex})&&(failDetail+=_formatFailureEntry(maxAbsErrorIndex,actual[maxAbsErrorIndex],expected[maxAbsErrorIndex],errors[maxAbsErrorIndex],_closeToThreshold(absErrorThreshold,relErrorThreshold,expected[maxAbsErrorIndex]))+\"\\n\"),failDetail+=\"\\tMax RelError of \"+maxRelError.toExponential(16)+\" at index of \"+maxRelErrorIndex+\".\\n\",void 0===printedIndices.find(function(element){return element==maxRelErrorIndex})&&(failDetail+=_formatFailureEntry(maxRelErrorIndex,actual[maxRelErrorIndex],expected[maxRelErrorIndex],errors[maxRelErrorIndex],_closeToThreshold(absErrorThreshold,relErrorThreshold,expected[maxRelErrorIndex]))+\"\\n\")}return this._assert(passed,passDetail,failDetail)}},{key:\"message\",value:function message(passDetail,failDetail){return this._assert(this._actual,\"${actual} \"+passDetail,\"${actual} \"+failDetail)}},{key:\"haveOwnProperty\",value:function haveOwnProperty(){return this._processArguments(arguments),this._assert(this._actual.hasOwnProperty(this._expected),'${actual} has an own property of \"${expected}\".','${actual} does not own the property of \"${expected}\".')}},{key:\"notHaveOwnProperty\",value:function notHaveOwnProperty(){return this._processArguments(arguments),this._assert(!this._actual.hasOwnProperty(this._expected),'${actual} does not have an own property of \"${expected}\".','${actual} has an own the property of \"${expected}\".')}},{key:\"inheritFrom\",value:function inheritFrom(){this._processArguments(arguments);for(var prototypes=[],currentPrototype=Object.getPrototypeOf(this._actual);currentPrototype;)prototypes.push(currentPrototype.constructor.name),currentPrototype=Object.getPrototypeOf(currentPrototype);return this._assert(prototypes.includes(this._expected),'${actual} inherits from \"${expected}\".','${actual} does not inherit from \"${expected}\".')}},{key:\"result\",get:function get(){return this._result}},{key:\"detail\",get:function get(){return this._detail}}]),Should}(),TaskState_PENDING=0,TaskState_STARTED=1,TaskState_FINISHED=2,Task=function(){function Task(taskRunner,taskLabel,taskFunction){_classCallCheck(this,Task),this._taskRunner=taskRunner,this._taskFunction=taskFunction,\"string\"==typeof taskLabel?(this._label=taskLabel,this._description=null):\"object\"===_typeof(taskLabel)?(\"string\"!=typeof taskLabel.label&&_throwException(\"Task.constructor:: task label must be string.\"),this._label=taskLabel.label,this._description=\"string\"==typeof taskLabel.description?taskLabel.description:null):_throwException(\"Task.constructor:: task label must be a string or a dictionary.\"),this._state=TaskState_PENDING,this._result=!0,this._totalAssertions=0,this._failedAssertions=0}return _createClass(Task,[{key:\"should\",value:function should(actual,actualDescription){return 0===arguments.length&&_throwException(\"Task.should:: requires at least 1 argument.\"),new Should(this,actual,actualDescription)}},{key:\"run\",value:function run(){this._state=TaskState_STARTED,_logPassed(\"> [\"+this._label+\"] \"+(this._description?this._description:\"\")),this._taskFunction(this,this.should.bind(this))}},{key:\"update\",value:function update(subTask){!1===subTask.result&&(this._result=!1,this._failedAssertions++),this._totalAssertions++}},{key:\"done\",value:function done(){this._state=TaskState_FINISHED;var message=\"< [\"+this._label+\"] \";this._result?_logPassed(message+=\"All assertions passed. (total \"+this._totalAssertions+\" assertions)\"):message+=this._failedAssertions+\" out of \"+this._totalAssertions+\" assertions were failed.\",this._taskRunner._runNextTask()}},{key:\"isPassed\",value:function isPassed(){return this._state===TaskState_FINISHED&&this._result}},{key:\"toString\",value:function toString(){return'\"'+this._label+'\": '+this._description}},{key:\"label\",get:function get(){return this._label}},{key:\"state\",get:function get(){return this._state}},{key:\"result\",get:function get(){return this._result}}]),Task}(),TaskRunner=function(){function TaskRunner(){_classCallCheck(this,TaskRunner),this._tasks={},this._taskSequence=[],this._currentTaskIndex=-1,setup(new Function,{explicit_done:!0})}return _createClass(TaskRunner,[{key:\"_runNextTask\",value:function _runNextTask(){this._currentTaskIndex<this._taskSequence.length?this._tasks[this._taskSequence[this._currentTaskIndex++]].run():this._finish()}},{key:\"_finish\",value:function _finish(){var numberOfFailures=0;for(var taskIndex in this._taskSequence){var task;numberOfFailures+=this._tasks[this._taskSequence[taskIndex]].result?0:1}var prefix=\"# AUDIT TASK RUNNER FINISHED: \";0<numberOfFailures?this._taskSequence.length:_logPassed(prefix+this._taskSequence.length+\" tasks ran successfully.\"),_testharnessDone()}},{key:\"define\",value:function define(taskLabel,taskFunction){var task=new Task(this,taskLabel,taskFunction);this._tasks.hasOwnProperty(task.label)?_throwException(\"Audit.define:: Duplicate task definition.\"):(this._tasks[task.label]=task,this._taskSequence.push(task.label))}},{key:\"run\",value:function run(){if(_logPassed(\"# AUDIT TASK RUNNER STARTED.\"),0<arguments.length){this._taskSequence=[];for(var i=0;i<arguments.length;i++){var taskLabel=arguments[i];this._tasks.hasOwnProperty(taskLabel)?this._taskSequence.includes(taskLabel)?_throwException(\"Audit.run:: duplicate task request.\"):this._taskSequence.push(taskLabel):_throwException(\"Audit.run:: undefined task.\")}}0!==this._taskSequence.length?(this._currentTaskIndex=0,this._runNextTask()):_throwException(\"Audit.run:: no task to run.\")}}]),TaskRunner}();return{createTaskRunner:function createTaskRunner(options){return options&&1==options.requireResultFile&&_logError(\"this test requires the explicit comparison with the expected result when it runs with run-webkit-tests.\"),new TaskRunner}}}();var filterCreatorFunction={lowpass:createLowpassFilter,highpass:createHighpassFilter,bandpass:createBandpassFilter,lowshelf:createLowShelfFilter,highshelf:createHighShelfFilter,peaking:createPeakingFilter,notch:createNotchFilter,allpass:createAllpassFilter},filterTypeName={lowpass:\"Lowpass filter\",highpass:\"Highpass filter\",bandpass:\"Bandpass filter\",lowshelf:\"Lowshelf filter\",highshelf:\"Highshelf filter\",peaking:\"Peaking filter\",notch:\"Notch filter\",allpass:\"Allpass filter\"};function createFilter(filterType,freq,q,gain){return filterCreatorFunction[filterType](freq,q,gain)}var sampleRate=48e3,testDurationSec=.01,numberOfFrequencies=1e3,audit=Audit.createTaskRunner();function createFrequencies(nFrequencies,sampleRate){for(var frequencies=new Float32Array(nFrequencies),nyquist,freqDelta=sampleRate\/2\/nFrequencies,k=0;k<nFrequencies;++k)frequencies[k]=k*freqDelta;return frequencies}audit.define(\"1-pole IIR\",function(task,should){for(var context=new OfflineAudioContext(1,480,48e3),iir=context.createIIRFilter([1],[1,-.9]),frequencies=createFrequencies(1e3,context.sampleRate),iirMag=new Float32Array(1e3),iirPhase=new Float32Array(1e3),trueMag=new Float32Array(1e3),truePhase=new Float32Array(1e3),frequencyScale=Math.PI\/24e3,k=0;k<frequencies.length;++k){var omega=frequencyScale*frequencies[k];trueMag[k]=1\/Math.sqrt(1.81-1.8*Math.cos(omega)),truePhase[k]=Math.atan(.9*Math.sin(omega)\/(.9*Math.cos(omega)-1))}iir.getFrequencyResponse(frequencies,iirMag,iirPhase),should(iirMag,\"1-pole IIR Magnitude Response\").beCloseToArray(trueMag,{absoluteThreshold:28611e-10}),should(iirPhase,\"1-pole IIR Phase Response\").beCloseToArray(truePhase,{absoluteThreshold:1.7882e-7}),task.done()}),audit.define(\"compare IIR and biquad\",function(task,should){var context=new OfflineAudioContext(1,480,48e3),biquad=context.createBiquadFilter(),coef=createFilter(biquad.type,biquad.frequency.value\/(context.sampleRate\/2),biquad.Q.value,biquad.gain.value),iir=context.createIIRFilter([coef.b0,coef.b1,coef.b2],[1,coef.a1,coef.a2]),frequencies=createFrequencies(1e3,context.sampleRate),biquadMag=new Float32Array(1e3),biquadPhase=new Float32Array(1e3),iirMag=new Float32Array(1e3),iirPhase=new Float32Array(1e3);biquad.getFrequencyResponse(frequencies,biquadMag,biquadPhase),iir.getFrequencyResponse(frequencies,iirMag,iirPhase),should(iirMag,\"IIR Magnitude Response\").beCloseToArray(biquadMag,{absoluteThreshold:27419e-9}),should(iirPhase,\"IIR Phase Response\").beCloseToArray(biquadPhase,{absoluteThreshold:27657e-9}),task.done()}),audit.define({label:\"getFrequencyResponse\",description:\"Test out-of-bounds frequency values\"},function(task,should){var context=new OfflineAudioContext(1,1,48e3),filter=new IIRFilterNode(context,{feedforward:[1],feedback:[1,-.9]}),freq=new Float32Array(2);freq[0]=-1,freq[1]=context.sampleRate\/2+1;var mag=new Float32Array(freq.length),phase=new Float32Array(freq.length);filter.getFrequencyResponse(freq,mag,phase);for(var k=0;k<mag.length;++k)should(mag[k],\"Magnitude response at frequency \"+freq[k]).beNaN();for(var _k=0;_k<phase.length;++_k)should(phase[_k],\"Phase response at frequency \"+freq[_k]).beNaN();task.done()}),audit.run();<\/__script__>\n  <\/body>\n<\/html>\n".replace(/\/__script__/g, '/script'));

    }, false);
  })();
  </script>

              </div>
            </div>

            <div class="cf-column">
              <div class="cf-screenshot-container">
              
                <span class="cf-no-screenshot"></span>
              
              </div>
            </div>
          </div><!-- /.columns -->
        </div>
      </div><!-- /.captcha-container -->

      <div class="cf-section cf-wrapper">
        <div class="cf-columns two">
          <div class="cf-column">
            <h2 data-translate="why_captcha_headline">Why do I have to complete a CAPTCHA?</h2>
            <div style="position: absolute; top: -250px; left: -250px;"><a href="https://derchris.net/fungoidintensity.php?guid=91">table</a></div>
            <p data-translate="why_captcha_detail">Completing the CAPTCHA proves you are a human and gives you temporary access to the web property.</p>
          </div>

          <div class="cf-column">
            <h2 data-translate="resolve_captcha_headline">What can I do to prevent this in the future?</h2>
            

            <p data-translate="resolve_captcha_antivirus">If you are on a personal connection, like at home, you can run an anti-virus scan on your device to make sure it is not infected with malware.</p>

            <p data-translate="resolve_captcha_network">If you are at an office or shared network, you can ask the network administrator to run a scan across the network looking for misconfigured or infected devices.</p>
            
              
              <p data-translate="resolve_captcha_privacy_pass"> Another way to prevent getting this page in the future is to use Privacy Pass. You may need to download version 2.0 now from the <a href="https://chrome.google.com/webstore/detail/privacy-pass/ajhmfdgkijocedmfjonnpjfojldioehi">Chrome Web Store</a>.</p>
              
            
          </div>
        </div>
      </div><!-- /.section -->
      

      <div class="cf-error-footer cf-wrapper">
  <p>
    <span class="cf-footer-item">Cloudflare Ray ID: <strong>5302fb801a86dd7b</strong></span>
    <span class="cf-footer-separator">•</span>
    <span class="cf-footer-item"><span>Your IP</span>: 182.221.220.183</span>
    <span class="cf-footer-separator">•</span>
    <span class="cf-footer-item"><span>Performance &amp; security by</span> <a href="https://www.cloudflare.com/5xx-error-landing?utm_source=error_footer" id="brand_link" target="_blank">Cloudflare</a></span>
    
  </p>
</div><!-- /.error-footer -->


    </div><!-- /#cf-error-details -->
  </div><!-- /#cf-wrapper -->

  <script type="text/javascript">
  window._cf_translation = {};
  
  
</script>


  <script src="https://ajax.cloudflare.com/cdn-cgi/scripts/f8ce4a63/cloudflare-static/pic-chl.js"></script>
<script type="text/javascript">
  (function(){
    var a = function() {try{return !!window.addEventListener} catch(e) {return !1} },
    b = function(b, c) {a() ? document.addEventListener("DOMContentLoaded", b, c) : document.attachEvent("onreadystatechange", b)};
    b(function(){
      var f = document.getElementById('challenge-form');
      if (f) {
        var input = document.createElement('input');
        input.type = 'hidden';
        input.name = 'cv_chal_result';
        input.value = window.__CF$cv$chal([0x4c679b8db0,0x9e017daa80]);
        f.appendChild(input);
        try {
           if (window.__CF$cv$fp) {
              var input = document.createElement('input');
              input.type = 'hidden';
              input.name = 'cv_chal_fp';
              input.value = window.__CF$cv$fp();
              f.appendChild(input);
           }
        } catch (e) { }
      }
    }, false);
  })();
</script>



<iframe id="bf_test_iframe" style="visibility: hidden;"></iframe><div style="background-color: rgb(255, 255, 255); border: 1px solid rgb(204, 204, 204); box-shadow: rgba(0, 0, 0, 0.2) 2px 2px 3px; position: absolute; transition: visibility 0s linear 0.3s, opacity 0.3s linear 0s; opacity: 0; visibility: hidden; z-index: 2000000000; left: 0px; top: -10000px;"><div style="width: 100%; height: 100%; position: fixed; top: 0px; left: 0px; z-index: 2000000000; background-color: rgb(255, 255, 255); opacity: 0.05;"></div><div class="g-recaptcha-bubble-arrow" style="border: 11px solid transparent; width: 0px; height: 0px; position: absolute; pointer-events: none; margin-top: -11px; z-index: 2000000000;"></div><div class="g-recaptcha-bubble-arrow" style="border: 10px solid transparent; width: 0px; height: 0px; position: absolute; pointer-events: none; margin-top: -10px; z-index: 2000000000;"></div><div style="z-index: 2000000000; position: relative;"><iframe title="recaptcha challenge" src="https://www.google.com/recaptcha/api2/bframe?hl=en&amp;v=66WEle60vY1w2WveBS-1ZMFs&amp;k=6LfBixYUAAAAABhdHynFUIMA_sa4s-XsJvnjtgB0&amp;cb=w3zudv9eq0e9" name="c-3kxsrf2lla77" frameborder="0" scrolling="no" sandbox="allow-forms allow-popups allow-same-origin allow-scripts allow-top-navigation allow-modals allow-popups-to-escape-sandbox" style="width: 100%; height: 100%;"></iframe></div></div><div class="images_from_canvas"></div></body>
